+++ src/main/java/com/exposure/exposureservice/service/SysRoleService.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface SysRoleService {
     List<SysRole> findAll();
     PageBean<List<SysRole>> findList(String roleName, Integer pageIndex, Integer pageSize);
-    SysRole findById(Integer id);
+    SysRole findById(Long id);
     void insert(SysRole sysRole);
     void update(SysRole sysRole);
-    void del(Integer id);
+    void del(Long id);
 }
